requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
docutils==0.19
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imagesize==1.4.1
imgkit==1.2.3
Jinja2==3.1.2
loguru @ file:///croot/loguru_1675318478402/work
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.2
OpenAIAuth==0.3.2
packaging==23.0
Pillow==9.4.0
prompt-toolkit==3.0.38
pydantic==1.10.7
Pygments==2.14.0
PySocks==1.7.1
python-markdown-math==0.8
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
revChatGPT==4.0.11
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
socksio==1.0.0
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starlette==0.26.1
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
